add rest of basic tests and correct issues
