+++ src/vigilant_server/__init__.py
@@ -1,3 +1,4 @@
-from .board import Board, Player
+from .board import Board, create_board
+from .player import Player
 from .card import Card
 from .node import Node
