+++ src/vigilant_server/units.py
@@ -15,5 +15,6 @@ def base_unit_card() -> Card:
     return Card(
         name="Base Card",
         desc="Base Desciption",
+        honey_cost=1,
         children=[Summoner(unit=base_unit, children=[])],
     )
